+++ unifier/apps/drf/v1/serializers/__init__.py
@@ -1,0 +1,12 @@
+from unifier.apps.drf.v1.serializers.manga import (
+    MangaChapterDetailSerializer,
+    MangaChapterSerializer,
+    MangaSerializer,
+    MangaSerializerDetail,
+)
+from unifier.apps.drf.v1.serializers.novel import (
+    NovelChapterDetailSerializer,
+    NovelChapterSerializer,
+    NovelSerializer,
+    NovelSerializerDetail,
+)
